requirements.txt
scrapy==2.4.0
requests==2.24.0
python-dateutil==2.8.1
lxml==4.6.1